Use bundling (#30)
weas and weas-widget are separate repos, but if we want to debug/develop a new feature of weas for the widget, using bundling makes the development more convenient.

The js code was moved to the `js` folder, run `npm run build-watch` will bundle these assets into the `src/weas_widget/static` folder, and watch the changes.

If we want to use the weas package, then clone the weas repo, and the two repo folders are in the same level.

If we want to use the release version from unpkg

- Use the release version
- run `npm run build`
- python -m build
- twine upload dist/*